xorg-gaming: Implement combination of Synergy + input flipping.
So I've been using the input flipping method for several months now and
it works really well. Never had a single technical problem with it.

However, pressing Ctrl+Alt+Tab all the time to flip back and forth can
get pretty annoying. That's especially true for games that don't have
any real need for the input to be locked to a particular screen.

I had a flash of brilliance: we can use both Synergy and the input
flipping method. With this setup, using Synergy to roam between screens
still works seamlessly. You can still lock the mouse to a particular
screen in Synergy using the pause key. Additionally, Ctrl+Alt+Tab is
still available to flip input over to the other X server if a
low-latency, pointer-locking solution is needed.

Why didn't I think of this before?